refac: moved constants to constants.ts
