+++ PancakeSwap-Bots/Large-swap/src/utils.ts
@@ -2,21 +2,9 @@ import { ethers, Finding, FindingType, FindingSeverity } from "forta-agent";
 import { createAddress } from "forta-agent-tools/lib/tests.utils";
 import BigNumber from "bignumber.js";
 import { getCreate2Address } from "@ethersproject/address";
+import { ERC20ABI, PANCAKE_PAIR_ABI } from "./constants";
 
 BigNumber.set({ DECIMAL_PLACES: 18 });
-
-const SWAP_EVENT =
-  "event Swap(address indexed sender, uint amount0In, uint amount1In, uint amount0Out,uint amount1Out,address indexed to)";
-
-const LARGE_THRESHOLD = "2.5"; // percent
-const ERC20ABI = ["function balanceOf(address account) public view returns (uint256)"];
-const PANCAKE_PAIR_ABI = [
-  "function token0() external view returns (address)",
-  "function token1() external view returns (address)",
-];
-const PANCAKE_FACTORY_ABI = ["function getPair(address tokenA, address tokenB) external view returns (address pair)"];
-const PANCAKE_FACTORY_ADDRESS = "0xcA143Ce32Fe78f1f7019d7d551a6402fC5350c73";
-const INIT_CODE_PAIR_HASH = "0x00fb7f630766e6a796048ea87d01acd3068e8ff67d078148a3fa3f4a84f69bd5";
 
 const toBn = (ethersBn: ethers.BigNumberish) => new BigNumber(ethersBn.toString());
 
@@ -33,9 +21,9 @@ const isValidPancakePair = async (
   const pairContract = createContract(pairAddress, PANCAKE_PAIR_ABI, provider);
   let token0Address: string, token1Address: string;
   try {
-      [token0Address, token1Address] = await Promise.all([
+    [token0Address, token1Address] = await Promise.all([
       pairContract.token0({ blockTag: block }),
-      pairContract.token1({ blockTag: block })
+      pairContract.token1({ blockTag: block }),
     ]);
   } catch (error) {
     return [false, "", ""];
@@ -100,17 +88,4 @@ const createFinding = (
   });
 };
 
-export {
-  createFinding,
-  getERC20Balance,
-  isValidPancakePair,
-  SWAP_EVENT,
-  LARGE_THRESHOLD,
-  ERC20ABI,
-  PANCAKE_PAIR_ABI,
-  toBn,
-  PANCAKE_FACTORY_ABI,
-  PANCAKE_FACTORY_ADDRESS,
-  INIT_CODE_PAIR_HASH,
-  getPancakePairCreate2Address,
-};
+export { createFinding, getERC20Balance, isValidPancakePair, toBn, getPancakePairCreate2Address };
